requirements.txt
click==8.1.3
colorama==0.4.6
dash==2.7.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
numpy==1.23.5
pandas==1.5.2
plotly==5.11.0
protobuf==3.20.1
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
tenacity==8.1.0
Werkzeug==2.2.2
xlrd==2.0.1